Deparse Row Expression as subquery
A row expression cannot be deparsed as ROW(...) cause otherwise column
names are lost when sending the deparsed query to the remote.

RowExpr may be present in the planner tree, as the optimizer may "pull
up" aggregation subqueries and it certainly does.

In particular, ST_AsMVT may produce different results or may miss the
geom column if column names are generated as f1, f2, ..., fn.

This solves that issue by deparsing the row expression as an alias and
embedding it into a subquery.

This is the intended syntax transformation in pseudo-code:

SELECT row_expr
FROM from_clause
WHERE where_caluse

  =>

SELECT alias
FROM (
  SELECT row_expr_fields
  FROM from_clause
  WHERE where_clause
) alias